backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.43
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
